requirements.txt
streamlit
langchain-community
langchain
langchain-huggingface
faiss-cpu
langchain-core
langchain-text-splitters
unstructured
pi-heif
numpy
tqdm
python-dotenv
langchain-ollama
pdfminer.six
unstructured[pdf]
langchain-openai
onnxruntime
sentence-transformers
